Fixing system tests and bug in master_application code for running LIST operation

+++ dss_datamover/tests/conftest.py
@@ -1,15 +1,45 @@
+#!/usr/bin/python3
 
-import os
-import sys
-from multiprocessing import Queue, Value, Lock
-import pytest
-
-top_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), os.pardir))
-sys.path.append(top_dir)
+"""
+# The Clear BSD License
+#
+# Copyright (c) 2023 Samsung Electronics Co., Ltd.
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted (subject to the limitations in the disclaimer
+# below) provided that the following conditions are met:
+#
+# * Redistributions of source code must retain the above copyright notice,
+#   this list of conditions and the following disclaimer.
+# * Redistributions in binary form must reproduce the above copyright notice,
+#   this list of conditions and the following disclaimer in the documentation
+#   and/or other materials provided with the distribution.
+# * Neither the name of Samsung Electronics Co., Ltd. nor the names of its
+#   contributors may be used to endorse or promote products derived from this
+#   software without specific prior written permission.
+# NO EXPRESS OR IMPLIED LICENSES TO ANY PARTY'S PATENT RIGHTS ARE GRANTED BY
+# THIS LICENSE. THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND
+# CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT
+# NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+# PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
+# CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+# EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+# PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+# OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+# WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
+# OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
+# ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+"""
 
 from utils import utility, config
 from logger import MultiprocessingLogger
-from master_application import Master, process_put_operation, process_list_operation, process_get_operation, process_del_operation  # noqa
+from master_application import Master, process_put_operation, process_list_operation, process_get_operation, process_del_operation
+from multiprocessing import Queue, Value, Lock
+
+import os
+import pytest
+
 
 @pytest.fixture(scope="session")
 def get_config_object():
@@ -25,13 +55,14 @@ def get_config_dict(get_config_object):
 
 @pytest.fixture(scope="session")
 def get_system_config_object():
-    config_obj = config.Config({})
+    config_obj = config.Config({}, config_filepath="/etc/dss/datamover/config.json")
     return config_obj
 
 
 @pytest.fixture(scope="session")
 def get_system_config_dict(get_system_config_object):
     return get_system_config_object.get_config()
+
 
 @pytest.fixture
 def get_multiprocessing_logger(tmpdir):
@@ -50,12 +81,13 @@ def get_multiprocessing_logger(tmpdir):
     # teardown logger
     logger.stop()
 
-# TODO: add utility to clear the cache, clear_datamover_cache
+
 @pytest.fixture
 def clear_datamover_cache():
     cache_files = ["/var/log/dss/prefix_index_data.json", "/var/log/dss/dm_resume_prefix_dir_keys.txt"]
     for f in cache_files:
-        os.remove(f)
+        if os.path.exists(f):
+            os.remove(f)
 
 
 class MockSocket():
@@ -66,22 +98,22 @@ class MockSocket():
     def __init__(self):
         self.mockhost = "192.168.300.144"
         self.mockport = "4000"
-    
+
     def connect(self):
         return True
-    
+
     def recv(self):
         pass
-    
+
     def sendall(self):
         pass
-    
 
-        
+
 @pytest.fixture
 def get_mock_clientsocket(mocker):
     mock_clientsocket = mocker.patch('socket_communication.ClientSocket', spec=True)
     return mock_clientsocket
+
 
 @pytest.fixture
 def get_mock_serversocket(mocker):
@@ -89,8 +121,11 @@ def get_mock_serversocket(mocker):
     return mock_serversocket
 
 @pytest.fixture
-def get_master(get_system_config_dict):
+def get_master_dryrun(get_system_config_dict):
     def instantiate_master_object(operation):
+        get_system_config_dict["config"] = "/etc/dss/datamover/config.json"
+        get_system_config_dict["dest_path"] = "/tmp/xyz"
+        get_system_config_dict["dryrun"] = True
         master = Master(operation, get_system_config_dict)
         print("instantiated master obj")
         master.start()
@@ -98,6 +133,28 @@ def get_master(get_system_config_dict):
         return master
     return instantiate_master_object
 
+
+@pytest.fixture
+def get_master(get_system_config_dict):
+    def instantiate_master_object(operation):
+        get_system_config_dict["config"] = "/etc/dss/datamover/config.json"
+        get_system_config_dict["dest_path"] = "/tmp/xyz"
+        master = Master(operation, get_system_config_dict)
+        print("instantiated master obj")
+        master.start()
+        print("successfully started master obj")
+        return master
+    return instantiate_master_object
+
+@pytest.fixture
+def shutdown_master_without_nfscluster():
+    def _method(master):
+        print("shutting down master")
+        master.stop_logging()
+        print("stopping logging")
+        master.stop_monitor()
+        print("stopping monitoring")
+    return _method
 
 @pytest.fixture
 def shutdown_master():
